Move mutex lock guard in its own, smaller header
